TEAMNADO-4959: Migrate to useChrome (#196)
Migrate to useChrome

This updates all instances of window.insights.chrome to useChrome(). It
touches basically everything because the token is retrieved from the
chroming. To combat this in the future, I have created a custom hook,
useToken so that if (when) the platform changes how this works we won't
have to change nearly as many things.