Teamnado 5960: Add manifest creation gate via chrome.auth.user.entitlements (#195)
* Add manifest creation gate via chrome.auth.user.entitlements

This adds a gate for manifest creation based on whether or not the user
has a satellite subscription. This is checked via whether or not the
user is entitled to smart_management.

When the user is entitled, the application functions the same as it has
previously. In the case that the user is _not entitled_, yet has
manifests, they should still be able to manage preexisting, but not
create new, manifests. In the case the user is _not entitled_ and has
_no manifests_ they should receive a friendly empty state that directs
them to support, and their subscriptions.

* Add tests

This adds unit testing for the manifest creation gate.